+++ platform/android/MapLibreAndroid/src/vulkan/java/org/maplibre/android/maps/renderer/surfaceview/VulkanSurfaceViewMapRenderer.java
@@ -8,7 +8,14 @@ public class VulkanSurfaceViewMapRenderer extends SurfaceViewMapRenderer {
   public VulkanSurfaceViewMapRenderer(Context context,
                                 @NonNull MapLibreVulkanSurfaceView surfaceView,
                                 String localIdeographFontFamily) {
-    super(context, surfaceView, localIdeographFontFamily);
+    this(context, surfaceView, localIdeographFontFamily, 1);
+  }
+
+  public VulkanSurfaceViewMapRenderer(Context context,
+                                @NonNull MapLibreVulkanSurfaceView surfaceView,
+                                String localIdeographFontFamily,
+                                int msaaSamples) {
+    super(context, surfaceView, localIdeographFontFamily, msaaSamples);
 
     this.surfaceView.setRenderer(this);
   }