+++ platform/glfw/glfw_gl_backend.cpp
@@ -93,7 +93,9 @@ namespace mbgl {
 namespace gfx {
 
 template <>
-std::unique_ptr<GLFWBackend> Backend::Create<mbgl::gfx::Backend::Type::OpenGL>(GLFWwindow* window, bool capFrameRate) {
+std::unique_ptr<GLFWBackend> Backend::Create<mbgl::gfx::Backend::Type::OpenGL>(GLFWwindow* window,
+                                                                               bool capFrameRate,
+                                                                               [[maybe_unused]] uint32_t msaaSamples) {
     MLN_TRACE_FUNC();
 
     return std::make_unique<GLFWGLBackend>(window, capFrameRate);